rename and move page components +
