+++ src/pages/Dashboard.tsx
@@ -1,12 +1,12 @@
-import HtmlToImageButton from "./HtmlToImageButton"
-import MySyntaxTrees from "./SyntaxTree/MySyntaxTrees"
-import TreeSettings from "./SyntaxTree/Settings/TreeSettings"
-import ThemeSettings from "./SyntaxTree/Theme/ThemeSettings"
+import HtmlToImageButton from "../components/HtmlToImageButton"
+import MySyntaxTrees from "../components/SyntaxTree/MySyntaxTrees"
+import TreeSettings from "../components/SyntaxTree/Settings/TreeSettings"
+import ThemeSettings from "../components/SyntaxTree/Theme/ThemeSettings"
 import { useState, useContext } from 'react'
-import SyntaxTreeContext from "./SyntaxTree/SyntaxTreeContext"
-import SyntaxTreeGenerator from "./SyntaxTree/SyntaxTreeGenerator"
-import SyntaxTreeCanvas from "./SyntaxTree/SyntaxTreeCanvas"
-import SyntaxTreeName from "./SyntaxTree/SyntaxTreeName"
+import SyntaxTreeContext from "../components/SyntaxTree/SyntaxTreeContext"
+import SyntaxTreeGenerator from "../components/SyntaxTree/SyntaxTreeGenerator"
+import SyntaxTreeCanvas from "../components/SyntaxTree/SyntaxTreeCanvas"
+import SyntaxTreeName from "../components/SyntaxTree/SyntaxTreeName"
 import CreateNewIcon from "../assets/createnew.svg?react"
 import SettingsIcon from "../assets/settings.svg?react"
 import MyTreesIcon from "../assets/mytrees.svg?react"
